+++ preprocess/baseline_correction.py
@@ -1,6 +1,5 @@
-from typing import Union, Optional
+from typing import Optional
 import numpy as np
-from module.ms_module import SpectrumBaseModule
 from logger import get_logger
 
 logger = get_logger("ms_preprocess")
@@ -66,97 +65,3 @@ def snip_baseline(y: np.ndarray,
         if rel_change < epsilon:
             break
     return y_work
-
-class MSIPreprocessor():
-    """
-    Abstract base class for MSI data preprocessing.
-    
-    This class provides a framework for implementing various preprocessing techniques
-    for Mass Spectrometry Imaging (MSI) data, including peak picking, normalization,
-    alignment, baseline correction, and noise reduction.
-    
-    Supports both traditional input (mz, msroi arrays) and MSI object input for
-    better integration with the MSI framework.
-    
-    Attributes:
-        msi_object (MSI, optional): MSI object containing data and metadata
-        preprocessing_params (dict): Parameters for preprocessing operations
-        processed_data (np.ndarray, optional): Processed MSI data
-    """
-
-    def __init__(self):
-        """
-        Initialize MSI preprocessor.
-        
-        Args:
-            msi_object (MSI, optional): MSI object containing data and metadata
-            preprocessing_params (dict, optional): Parameters for preprocessing operations
-            
-        Raises:
-            ValueError: If neither msi_object nor (mz, msroi) are provided
-        """
-
-    @staticmethod
-    def baseline_correction(
-        data: Union[np.ndarray, SpectrumBaseModule],
-        method: str = "asls",
-        lam: float = 1e7,
-        p: float = 0.01,
-        niter: int = 15,
-        baseline_scale: float = 0.8,
-        m: Optional[int] = None,
-        decreasing: bool = True,
-        epsilon: float = 1e-3
-    ) -> tuple[Union[np.ndarray, SpectrumBaseModule], np.ndarray]:
-        """
-        Remove baseline drift and background signals from mass spectra.
-
-        Supports:
-        - ASLS (asymmetric least squares): robust baseline estimation with peak preservation.
-        - SNIP (statistics-sensitive non-linear iterative peak-clipping): progressive clipping with adaptive early-stop.
-
-        Args:
-            data: intensity array (np.ndarray) or SpectrumBaseModule
-            method: 'asls' (default) or 'snip'
-            lam: ASLS smoothness parameter (1e4-1e8, higher = smoother baseline)
-            p: ASLS asymmetry parameter (0.001-0.1, lower = more peak preservation)
-            niter: ASLS iterations (5-20)
-            baseline_scale: scale factor (0-1) applied to the estimated baseline
-            m: SNIP max half-window; None -> auto(min(50, n//10))
-            decreasing: SNIP iteration order; True: p=m..1 (coarse->fine), False: p=1..m
-            epsilon: SNIP adaptive early-stop threshold on relative change
-
-        Returns:
-            A tuple containing:
-            - np.ndarray or SpectrumBaseModule with baseline-corrected intensity.
-            - np.ndarray: The estimated baseline.
-        """
-        
-        def apply_single(intensity: np.ndarray):
-            xi = np.array(intensity, dtype=np.float64, copy=True)
-            xi = np.ascontiguousarray(xi)
-            # Estimate baseline via module-level functions
-            if method == "asls":
-                baseline = asls_baseline(xi, lam=lam, p=p, niter=niter)
-            elif method == "snip":
-                baseline = snip_baseline(xi, m=m, decreasing=decreasing, epsilon=epsilon)
-            else:
-                raise ValueError("Unsupported baseline method: use 'asls' or 'snip'")
-            scale = float(np.clip(baseline_scale, 0.0, 1.0))
-            scaled_baseline = scale * baseline
-            corrected = xi - scaled_baseline
-            corrected = np.maximum(corrected, 0.0)
-            return corrected, scaled_baseline
-
-        if isinstance(data, np.ndarray):
-            return apply_single(data)
-        elif isinstance(data, SpectrumBaseModule):
-            corrected_intensity, baseline = apply_single(np.array(data.intensity, dtype=np.float64))
-            corrected_spectrum = SpectrumBaseModule(
-                mz_list=np.array(data.mz_list, dtype=np.float64, copy=True) if data.mz_list is not None else None,
-                intensity=np.ascontiguousarray(corrected_intensity),
-                coordinates=data.coordinates
-            )
-            return corrected_spectrum, baseline
-        else:
-            raise TypeError("data must be np.ndarray or SpectrumBaseModule")
